requirements.txt
discord~=2.1.0
schedule~=1.1.0
python-dotenv~=0.21.0
requests~=2.28.1